--- backend/src/controller/vacina.controller.js
@@ -1,21 +1,0 @@
-const {getVacineByPeriodService, getVaccineProtectionService} = require("../service/vacinaService")
-
-const getVacineByPeriodController = async(req, res) => {
-    const response = await getVacineByPeriodService()
-
-    return res.status(200).json(response.rows)
-
-}
-
-const getVaccineProtectionController = async(req, res) => {
-    const disease = req.params.doenca
-    const response = await getVaccineProtectionService(disease)
-
-    return res.status(200).json(response.rows)
-
-}
-
-module.exports = {
-    getVacineByPeriodController,
-    getVaccineProtectionController
-}
