+++ backend/src/service/vacinneByPeriodService.js
@@ -14,15 +14,8 @@ const getVacineByPeriodService = async() => {
     
 }
 
-const getVaccineProtectionService = async(disease) => {
-    disease = "%" + disease + "%"
-    const response = await pool.query(`SELECT * FROM vacina
-    WHERE doenca_protecao ILIKE $1`, [disease])
 
-    return response
-}
 
 module.exports = {
-    getVacineByPeriodService,
-    getVaccineProtectionService
+    getVacineByPeriodService
 }
